+++ driver-core/CHANGELOG.rst
@@ -23,6 +23,8 @@ Merged from 2.0 branch:
   (JAVA-593)
 - [bug] Ensure state change notifications for a given host are handled serially
   (JAVA-594)
+- [bug] Ensure control connection reconnects when control host is removed
+  (JAVA-597)
 
 
 2.1.3:
@@ -167,6 +169,8 @@ Merged from 2.0 branch: everything up to 2.0.3 (included), and the following.
   (JAVA-593)
 - [bug] Ensure state change notifications for a given host are handled serially
   (JAVA-594)
+- [bug] Ensure control connection reconnects when control host is removed
+  (JAVA-597)
 
 
 2.0.8: